+++ sema4ai/src/sema4ai_code/commands.py
@@ -139,6 +139,8 @@ SEMA4AI_GET_RUNBOOK_PATH_FROM_AGENT_SPEC_INTERNAL = "sema4ai.getRunbookPathFromA
 SEMA4AI_AGENT_PACKAGE_PUBLISH_TO_SEMA4_AI_STUDIO_APP = "sema4ai.agentPackagePublishToSema4AIStudioApp"  # Publish Agent Package to Sema4.ai Studio
 SEMA4AI_AGENT_PACKAGE_IMPORT = "sema4ai.agentPackageImport"  # Import Agent Package (Zip)
 SEMA4AI_REFRESH_AGENT_SPEC_INTERNAL = "sema4ai.refreshAgentSpec.internal"  # Refresh Agent Spec (internal)
+SEMA4AI_UPDATE_AGENT_VERSION = "sema4ai.updateAgentVersion"  # Update Agent Version
+SEMA4AI_UPDATE_AGENT_VERSION_INTERNAL = "sema4ai.updateAgentVersion.internal"  # Update Agent Version (internal)
 
 ALL_SERVER_COMMANDS = [
     SEMA4AI_GET_PLUGINS_DIR,
@@ -195,6 +197,7 @@ ALL_SERVER_COMMANDS = [
     SEMA4AI_PACK_AGENT_PACKAGE_INTERNAL,
     SEMA4AI_GET_RUNBOOK_PATH_FROM_AGENT_SPEC_INTERNAL,
     SEMA4AI_REFRESH_AGENT_SPEC_INTERNAL,
+    SEMA4AI_UPDATE_AGENT_VERSION_INTERNAL,
 ]
 
 # fmt: on